CAPT-218 log to secondary when can't read/write from queue
